requirements.txt
fastapi
uvicorn
requests
python-multipart
jinja2
psycopg2-binary
